+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: aachhoub <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 14:39:48 by aachhoub          #+#    #+#             */
-/*   Updated: 2022/10/07 14:07:22 by aachhoub         ###   ########.fr       */
+/*   Updated: 2022/10/18 09:53:24 by aachhoub         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 	while (i < n)
 	{
 		if (b1[i] != b2[i])
-			return ((int)(b1[i]) - (int)(b2[i]));
+			return (b1[i] - b2[i]);
 		i++;
 	}
 	return (0);
